requirements/base.txt
# Python
# ------------------------------------------------------------------------------
requests==2.28.2
psycopg2-binary==2.9.10
Pillow==10.3.0
argon2-cffi==23.1.0
# Django
# ------------------------------------------------------------------------------
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-mptt==0.16.0
django-redis==5.4.0
# Django REST Framework
# ------------------------------------------------------------------------------
djangorestframework==3.15.2
djangorestframework-camel-case==1.4.2
djangorestframework-simplejwt==5.3.1
drf-nested-forms==1.1.8
drf-spectacular==0.28.0
# Code quality
# ------------------------------------------------------------------------------
ruff==0.4.2
coverage==7.5.0
djlint==1.34.1
pre-commit==3.7.0
# Async
# ------------------------------------------------------------------------------
celery==5.4.0
django-celery-beat==2.7.0
django-celery-results==2.5.1
flower==2.0.1
